Refactor: data-loading -> data
