+++ src/features/data/loadObjectsFromFile.ts
