+++ src/features/data/population/computeLocaleWritingPopulation.ts
