+++ src/features/transforms/filtering/WritingSystemFilterSelector.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback } from 'react';
 
-import { useDataContext } from '@features/data-loading/context/useDataContext';
+import { useDataContext } from '@features/data/context/useDataContext';
 import { PageParamKey, SearchableField } from '@features/params/PageParamTypes';
 import {
   SelectorDisplay,
