+++ src/pages/DataPage.tsx
@@ -5,7 +5,7 @@ import Loading from '@widgets/Loading';
 import HoverCardProvider from '@features/hovercard/HoverCardProvider';
 
 const PageParamsProvider = React.lazy(() => import('@features/params/PageParamsProvider'));
-const DataProvider = React.lazy(() => import('@features/data-loading/context/DataProvider'));
+const DataProvider = React.lazy(() => import('@features/data/context/DataProvider'));
 const DataPageBody = React.lazy(() => import('./DataPageBody'));
 const SidePanel = React.lazy(() => import('@widgets/controls/SidePanel'));
 const ViewModal = React.lazy(() => import('@features/modal/ViewModal'));
